+++ stroom-pipeline/src/main/java/stroom/pipeline/refdata/ReferenceDataServiceImpl.java
@@ -1,5 +1,6 @@
 package stroom.pipeline.refdata;
 
+import stroom.bytebuffer.ByteBufferPool;
 import stroom.dashboard.expression.v1.Val;
 import stroom.dashboard.expression.v1.ValInteger;
 import stroom.dashboard.expression.v1.ValLong;
@@ -35,6 +36,7 @@ import stroom.query.api.v2.ExpressionTerm.Condition;
 import stroom.query.common.v2.DateExpressionParser;
 import stroom.security.api.SecurityContext;
 import stroom.security.shared.PermissionNames;
+import stroom.task.api.TaskContext;
 import stroom.task.api.TaskContextFactory;
 import stroom.util.logging.LambdaLogger;
 import stroom.util.logging.LambdaLoggerFactory;
@@ -184,6 +186,7 @@ public class ReferenceDataServiceImpl implements ReferenceDataService {
     private final PipelineScopeRunnable pipelineScopeRunnable;
     private final TaskContextFactory taskContextFactory;
     private final RefDataValueProxyConsumerFactory.Factory refDataValueProxyConsumerFactoryFactory;
+    private final ByteBufferPool byteBufferPool;
 
     @Inject
     public ReferenceDataServiceImpl(final RefDataStoreFactory refDataStoreFactory,
@@ -193,7 +196,8 @@ public class ReferenceDataServiceImpl implements ReferenceDataService {
                                     final RefDataValueConverter refDataValueConverter,
                                     final PipelineScopeRunnable pipelineScopeRunnable,
                                     final TaskContextFactory taskContextFactory,
-                                    final Factory refDataValueProxyConsumerFactoryFactory) {
+                                    final Factory refDataValueProxyConsumerFactoryFactory,
+                                    final ByteBufferPool byteBufferPool) {
         this.refDataStore = refDataStoreFactory.getOffHeapStore();
         this.securityContext = securityContext;
         this.feedStore = feedStore;
@@ -202,6 +206,7 @@ public class ReferenceDataServiceImpl implements ReferenceDataService {
         this.pipelineScopeRunnable = pipelineScopeRunnable;
         this.taskContextFactory = taskContextFactory;
         this.refDataValueProxyConsumerFactoryFactory = refDataValueProxyConsumerFactoryFactory;
+        this.byteBufferPool = byteBufferPool;
     }
 
     @Override
@@ -315,6 +320,11 @@ public class ReferenceDataServiceImpl implements ReferenceDataService {
                                 () -> refDataStore.purge(refStreamId, partIndex),
                                 LogUtil.message("Performing Purge for ref stream {}:{}", refStreamId, partIndex)))
                         .run());
+    }
+
+    @Override
+    public void clearBufferPool() {
+        byteBufferPool.clear();
     }
 
     private void performPurge(final StroomDuration purgeAge) {
@@ -456,13 +466,16 @@ public class ReferenceDataServiceImpl implements ReferenceDataService {
                        final Consumer<Val[]> consumer) {
 
         withPermissionCheck(() -> LOGGER.logDurationIfInfoEnabled(
-                () -> doSearch(criteria, fields, consumer),
+                () -> taskContextFactory.context("Querying reference data store", taskContext ->
+                        doSearch(criteria, fields, consumer, taskContext))
+                        .run(),
                 "Querying ref store"));
     }
 
     private void doSearch(final ExpressionCriteria criteria,
                           final AbstractField[] fields,
-                          final Consumer<Val[]> consumer) {
+                          final Consumer<Val[]> consumer,
+                          final TaskContext taskContext) {
         // TODO @AT This is a temporary very crude impl to see if it works.
         //  The search code ought to be pushed down to the offHeapStore so it can query the many DBs
         //  selectively and in a MUCH more efficient way, e.g. using start/stop keys on the kv store scan.
@@ -500,7 +513,17 @@ public class ReferenceDataServiceImpl implements ReferenceDataService {
                     .skip(skipCount)
                     .limit(limit)
                     .forEach(refStoreEntry -> {
+                        if (taskContext.isTerminated()) {
+                            throw new RuntimeException("Aborting search due to task termination");
+                        }
+
                         final Val[] valArr = new Val[fields.length];
+
+                        try {
+                            Thread.sleep(50);
+                        } catch (InterruptedException e) {
+                            Thread.currentThread().interrupt();
+                        }
 
                         for (int i = 0; i < fields.length; i++) {
                             AbstractField field = fields[i];